requirements.txt
Flask
Flask-CORS
pandas
google-generativeai
python-dotenv